requirements.txt

fastapi==0.115.2
fastapi-mcp==0.2.0
openai-whisper==20240930

spacy==3.8.2
streamlit==1.39.0
aiohttp==3.10.10
pydantic==2.9.2
sounddevice==0.5.0
numpy==1.26.4
scipy==1.14.1
uvicorn==0.32.0
python-dotenv==1.0.1